Adding readonly to the disabled styles…
Use case:
I have a form. When the user clicks on a (fancy) submit it gets a disabled attribute. Now the button's value doesn't get submitted and the server has no way of knowing which button the user clicked. More useful is setting the button to readonly - avoids multiple form submissions but still sends its value.